fix: separate tests to improve readability
