+++ x/alliance/keeper/tests/reward_test.go
@@ -1,4 +1,4 @@
-package keeper_test
+package tests_test
 
 import (
 	"testing"
