+++ x/alliance/types/tests/types_test.go
@@ -1,4 +1,4 @@
-package types_test
+package tests_test
 
 import (
 	"testing"
